+++ docs/changes.rst
@@ -14,6 +14,7 @@ Version 0.5.7
 
 Unreleased.
 
+ - Added :meth:`Image.contrast() <wand.image.BaseImage.contrast>` method.
  - Added :meth:`Image.local_contrast() <wand.image.BaseImage.local_contrast>` method.
  - Added :meth:`Image.ordered_dither() <wand.image.BaseImage.ordered_dither>` method.
  - Added :meth:`Image.random_threshold() <wand.image.BaseImage.random_threshold>` method.